Moves data generation to separate file
